Better dictionary inspector, RenderAPI viewport is not in normalized coords instead of screenspace

+++ Crowny/Source/Crowny/RenderAPI/RenderAPI.h
@@ -32,7 +32,7 @@ namespace Crowny
 
         virtual void SubmitCommandBuffer(const Ref<CommandBuffer>& commandBuffer, uint32_t syncMask = 0xFFFFFFFF) = 0;
 
-        virtual void SetViewport(uint32_t x, uint32_t y, uint32_t width, uint32_t height,
+        virtual void SetViewport(float x, float y, float width, float height,
                                  const Ref<CommandBuffer>& commandBuffer = nullptr) = 0;
         virtual void SetClearColor(
           const glm::vec4& color) = 0; // TODO: Replace, const Ref<CommandBuffer>& commandBuffer = nullptr with color
